Fix a few tab completion bugs
When there are more than two matches, allow cycling through them.
Previously it just alternated between the first two.

When tabbing from a different place, start the cycle back from
the start instead of continuing.

When there's only one result tabbing again should do nothing.
Previously it returned to the original input. (There was a test
for this already that was failing)